Added the ability to vary fluid paths. Important for testing a variety of designs. More work needed to fully implement these. See TODOs in code.
Changed optimise class to Optimise_Worker which is a QRunnable that can be run in a QThreadPool

Added some preliminary restraint functions to be completed.
calculate mass, check if the design is geometricly feasible. This is quite a hard one as there are lots of
manufacturering constraints. We also need a way to position the pipe strucutre